+ move Deal.java to /server/deal
