--- src/main/java/lubiezurek/texasholdem/server/Deal.java
@@ -1,18 +1,0 @@
-package lubiezurek.texasholdem.server.deal;
-
-import states.IState;
-
-public class Deal{
-	IState currentState;
-
-	Deal(){
-
-	}
-
-	IState getState(){
-		return currentState;
-	}
-
-
-
-}
