+++ src/main/java/lubiezurek/texasholdem/server/deal/Deal.java
@@ -1,4 +1,6 @@
-package lubiezurek.texasholdem.server;
+package lubiezurek.texasholdem.server.deal;
+
+import states.IState;
 
 public class Deal{
 	IState currentState;
